Add files via upload
In the Bubble Sort algorithm. The outer loop (for(int i=0; i<a.length; i++)) iterates over each element in the array. 
The inner loop (for(int j=0; j<a.length-1; j++)) compares adjacent elements and swaps them if they are in the wrong order. 
This process is repeated until the entire array is sorted.

The condition if (a[j] > a[j + 1]) checks if the current element is greater than the next one. 
If true, a swap is performed using the temporary variable temp. 
This causes the larger element to "bubble up" to its correct position in the sorted order.