Add files via upload
The Selection Sort algorithm. The outer loop (for(int i=0; i<a.length; i++)) iterates over each element in the array. Within this loop:

min is initially set to the index i, representing the assumed minimum element in the unsorted portion of the array.
The inner loop (for(int j=i+1; j<a.length; j++)) iterates over the remaining unsorted elements, 
and if a smaller element is found, the index min is updated to the index of the smaller element.
After finding the minimum element in the unsorted portion, a swap is performed between the current element at index i and 
the minimum element found (a[min]).
This process is repeated until the entire array is sorted.